the processing is divided into pre- and postprocessing with two different operators.
meanReflPerTile is changed into a map of 2d-arrays: the integer-index gives the number of the tile, the array has the dimension 3x length(cloud path).
the 3 varieties cover 1. statistics from all pixel, 2. from land pixel, 3. from water pixel (the most reliable one for the determination of the best shift is the first one)

Input for the "postprocessing" are the original products, the preprocessing product is not used. apart from getting the per tile information to estimate the best shift, the processing is done in "postprocessing" (which should therefore be considered as processing)

findOverallMinimumReflectance estimates the best shift in the cloud mask. the method has been refined, it uses relative minima, which are not the first or the last entry along the cloud path.