(20/24) fix for NPW was already on master original commit: https://github.com/bcdev/snap-idepix/pull/76/commits/640007f46cfdd673d3aef4493f18311eb2f0ea79
additional changes:
rectCalculator was always instantiated with default  cloud buffer width or what was given as parameter.
This leads to additional computation. Now, creating an rectCalculator (rectExtender) depending on the needs.
With no extent, cloud buffer extent or extent for cloud shadow.
Also added validation of correct configuration of computeCloudShadow and ctpProduct and ctpBand